re-organize clustering folder
