+++ src/spikeinterface/sortingcomponents/clustering/main.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from .method_list import *
-
 from spikeinterface.core.job_tools import fix_job_kwargs, _shared_job_kwargs_doc
 
 import copy
