+++ src/spikeinterface/sortingcomponents/clustering/tests/test_merging_tools.py
@@ -1,0 +1,12 @@
+import pytest
+import numpy as np
+
+from spikeinterface.sortingcomponents.clustering.merging_tools import merge_peak_labels_from_features, merge_peak_labels_from_templates
+
+# TODO
+def test_merge():
+    pass
+
+
+if __name__ == "__main__":
+    test_merge()
