+++ src/spikeinterface/sortingcomponents/clustering/tests/test_splitting_tools.py
@@ -1,0 +1,12 @@
+import pytest
+import numpy as np
+
+from spikeinterface.sortingcomponents.clustering.splitting_tools import split_clusters
+
+# T ODO
+def test_split():
+    pass
+
+
+if __name__ == "__main__":
+    test_split()
