--- src/spikeinterface/sortingcomponents/tests/test_split.py
@@ -1,14 +1,0 @@
-import pytest
-import numpy as np
-
-from spikeinterface.sortingcomponents.clustering.split import split_clusters
-
-# no proper test at the moment this is used in tridesclous2
-
-
-def test_split():
-    pass
-
-
-if __name__ == "__main__":
-    test_split()
